+++ packages/ide-extension/src/commands/extractMessage.ts
@@ -20,7 +20,7 @@ export const extractMessageCommand = {
 		// guards
 		if (!ideExtension) {
 			return msg(
-				"There is no `plugin` configuration for the inlang extension. One of the `modules` should expose a `plugin` which has `customApi` containing `app.inlang.ideExtension`",
+				"There is no `plugin` configuration for the Visual Studio Code extension (Sherlock). One of the `modules` should expose a `plugin` which has `customApi` containing `app.inlang.ideExtension`",
 				"warn",
 				"notification"
 			)