+++ packages/paraglide/paraglide-js-adapter-next/examples/pages/src/pages/_document.tsx
@@ -1,0 +1,14 @@
+import { languageTag } from "@/paraglide/runtime"
+import { Html, Head, Main, NextScript } from "next/document"
+
+export default function Document() {
+	return (
+		<Html lang={languageTag()}>
+			<Head />
+			<body>
+				<Main />
+				<NextScript />
+			</body>
+		</Html>
+	)
+}