+++ packages/versioned-interfaces/plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/plugin
 
+## 2.4.6
+
+### Patch Changes
+
+- 960f8fb70: rename the vscode extension to "Sherlock"
+
 ## 2.4.5
 
 ### Patch Changes